Moved AnimePage to Fit Organization
Moved AnimePage with the rest of the pages for organization etc.
